# [Performance Improvement] Add custom bulk scorer for hybrid query (2-3x faster)