… CHANGELOG.md
… qa/rolling-upgrade/src/test/java/org/opensearch/neuralsearch/bwc/rolling/HybridSearchRelevancyIT.java
… src/main/java/org/opensearch/neuralsearch/search/collector/HybridTopFieldDocSortCollector.java
… src/main/java/org/opensearch/neuralsearch/search/collector/HybridTopScoreDocCollector.java
… src/main/java/org/opensearch/neuralsearch/search/collector/PagingFieldCollector.java
… src/main/java/org/opensearch/neuralsearch/search/collector/SimpleFieldCollector.java
… src/test/java/org/opensearch/neuralsearch/query/HybridBulkScorerTests.java
… src/test/java/org/opensearch/neuralsearch/query/HybridQueryDocIdStreamTests.java
… src/test/java/org/opensearch/neuralsearch/query/HybridQueryScorerTests.java
… src/test/java/org/opensearch/neuralsearch/query/HybridScorerSupplierTests.java
… src/test/java/org/opensearch/neuralsearch/query/HybridSubQueryScorerTests.java
… src/test/java/org/opensearch/neuralsearch/search/collector/HybridCollectorTestCase.java
… src/test/java/org/opensearch/neuralsearch/search/collector/HybridTopFieldDocSortCollectorTests.java
… src/test/java/org/opensearch/neuralsearch/search/collector/HybridTopScoreDocCollectorTests.java
… src/test/java/org/opensearch/neuralsearch/search/query/HybridCollectorManagerTests.java
… src/testFixtures/java/org/opensearch/neuralsearch/BaseNeuralSearchIT.java