### Description
Adding custom implementation of bulk scorer for hybrid query, mainly design that is highlighted in this [RFC](https://github.com/opensearch-project/neural-search/issues/1290). This improves score collection process and overall response times for query. 

After running benchmarks using released 3.0 as baseline these are following results in terms of response time and system throughput:

<h4>NOAA dataset, keyword search)</h4>

https://github.com/opensearch-project/opensearch-benchmark-workloads/tree/main/noaa_semantic_search

**Benchmark Mode Performance**

Query Type | New solution (ops/s) | Base 3.0 (ops/s) | Improvement
-- | -- | -- | --
Simple Range | 91.2 | 93.8 | -2.80%
Medium Subset | 91.2 | 70 | 30.30%
Large Subset | 60.2 | 28 | 115%
3 Sub-queries | 39.6 | 11.2 | 253.60%

**Response times**

Query Type | Metric | New solution (ms) | Base 3.0 (ms) | Improvement
-- | -- | -- | -- | --
Medium Subset | p50 | 81.2 | 107.7 | 24.6%
  | p90 | 86.6 | 116.6 | 25.7%
  | p99 | 92.5 | 128.5 | 28.0%
Large Subset | p50 | 126.6 | 278.7 | 54.6%
  | p90 | 132.9 | 304.7 | 56.4%
  | p99 | 141.1 | 337.8 | 58.2%
3 Sub-queries | p50 | 184.5 | 656.7 | 71.9%
  | p90 | 214.9 | 880.5 | 75.6%
  | p99 | 231.6 | 1042.6 | 77.8%

<h4>Quora dataset, keyword + semantic search</h4>

Based on [neural_search](https://github.com/opensearch-project/opensearch-benchmark-workloads/tree/main/neural_search) and [PR](https://github.com/opensearch-project/opensearch-benchmark-workloads/pull/624) 

**Benchmark Mode Performance**

Query Type | New solution (ops/s) | Base 3.0 (ops/s) | Improvement
-- | -- | -- | --
Hybrid Search | 41.3 | 38.8 | 6.40%
Complex Hybrid | 34.9 | 33.2 | 5.10%
Complex with Aggs | 31.2 | 26.4 | 18.20%

**Response Time Metrics**

Query Type | Metric | New solution (ms) | Base 3.0 (ms) | Improvement
-- | -- | -- | -- | --
Hybrid Search | p50 | 179.9 | 197.1 | 8.70%
  | p90 | 202.4 | 224.9 | 10%
  | p99 | 229.2 | 251.1 | 8.70%
Complex Hybrid | p50 | 210.6 | 231.5 | 9%
  | p90 | 248.2 | 282.9 | 12.30%
  | p99 | 285.5 | 329.4 | 13.30%
Complex with Aggs | p50 | 217.7 | 291 | 25.20%
  | p90 | 254.9 | 379.6 | 32.90%
  | p99 | 292.4 | 448.8 | 34.80%

Cluster setup: 
- 3 data nodes cluster
- C5.2xlarge instance type (8 vCPU, 16Gb RAM)
- 6 shards
- 8 segments per shard, force merged before running search queries

### Related Issues
https://github.com/opensearch-project/neural-search/issues/1234
https://github.com/opensearch-project/neural-search/issues/1236 (partially)

### Check List
- [X] New functionality includes testing.
- ~~[ ] New functionality has been documented.~~
- ~~[ ] API changes companion pull request [created](https://github.com/opensearch-project/opensearch-api-specification/blob/main/DEVELOPER_GUIDE.md).~~
- [X] Commits are signed per the DCO using `--signoff`.
- [X] Public documentation issue/PR [created](https://github.com/opensearch-project/documentation-website/issues/new/choose).

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.
For more information on following Developer Certificate of Origin and signing off your commits, please check [here](https://github.com/opensearch-project/neural-search/blob/main/CONTRIBUTING.md#developer-certificate-of-origin).